Refactor migration for unit_amount column in sale_details table

+++ database/migrations/1773000020006_sale_details_add_unit_amount.ts
@@ -1,17 +1,19 @@
 import { BaseSchema } from '@adonisjs/lucid/schema'
 
 export default class AddUnitAmountToSaleDetails extends BaseSchema {
-    protected tableName = 'sale_details'
+  protected tableName = 'sale_details'
 
-    public async up() {
-        this.schema.alterTable(this.tableName, (table) => {
-            table.decimal('unit_amount', 15, 2).notNullable().defaultTo(0)
-        })
-    }
+  public async up() {
+    if (!(await this.schema.hasColumn(this.tableName, 'unit_amount')))
+      this.schema.alterTable(this.tableName, (table) => {
+        table.decimal('unit_amount', 15, 2).notNullable().defaultTo(0)
+      })
+  }
 
-    public async down() {
-        this.schema.alterTable(this.tableName, (table) => {
-            table.dropColumn('unit_amount')
-        })
-    }
+  public async down() {
+    if (await this.schema.hasColumn(this.tableName, 'unit_amount'))
+      this.schema.alterTable(this.tableName, (table) => {
+        table.dropColumn('unit_amount')
+      })
+  }
 }
